Move content out of v1 directory
