--- nightfall/__init__.py
@@ -1,10 +1,0 @@
-"""
-nightfall module
-~~~~~~~~~~~~~~~~
-
-    This module provides an SDK for Nightfall.
-
-    :copyright: (c) 2021 Nightfall
-    :license: MIT, see LICENSE for more details.
-"""
-from .api import Nightfall
