--- tests/mocks/mock_single_finding_response
@@ -1,1 +1,0 @@
-{"fragment": "th1Sisaf4k34p1k3y", "detectorName": "API key", "confidence": "LIKELY", "location": {"byteRange": {"start": 86, "end": 103}, "unicodeRange": {"start": 86, "end": 103}}}
