--- v1/dev-requirements.txt
@@ -1,3 +1,0 @@
-pytest
-pytest-cov
-requests
